Update README.md
Author here was trying to show how the string formatting works with the minimum amount of characters and the answer here only shows the numbers produced which misses the point. Just needed a few spaces here and there. My first proposed change in GitHub excuse me if I misunderstood something.